+++ app/src/main/java/com/cornellappdev/scoop/ui/onboarding/OnboardingProfile.kt
@@ -26,8 +26,7 @@ import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.unit.dp
 import com.cornellappdev.scoop.R
 import com.cornellappdev.scoop.onboarding.OnboardingFooter
-import com.cornellappdev.scoop.onboarding.OnboardingHeader
-import com.cornellappdev.scoop.ui.theme.DarkGray
+import com.cornellappdev.scoop.onboarding.NavHeader
 import com.cornellappdev.scoop.ui.theme.DarkGreen
 import com.cornellappdev.scoop.ui.theme.Gray
 import com.cornellappdev.scoop.ui.theme.Green
@@ -43,7 +42,7 @@ fun OnboardingProfile(pagerState: PagerState) {
         Modifier.background(Color.White)
     ) {
 
-        OnboardingHeader(pagerState = pagerState, title = "Profile", hasBackArrow = true)
+        NavHeader(pagerState = pagerState, title = "Profile", hasBackArrow = true)
 
         Column(
             modifier = Modifier
