+++ app/src/main/java/com/cornellappdev/scoop/ui/onboarding/OnboardingView5.kt
@@ -2,20 +2,15 @@ package com.cornellappdev.scoop.onboarding
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.Text
 import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.cornellappdev.scoop.components.BackArrow
 import com.cornellappdev.scoop.components.RightArrow
 import com.cornellappdev.scoop.ui.components.general.DenseTextField
-import com.cornellappdev.scoop.ui.components.general.UnderlinedEditText
 import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.PagerState
 
@@ -33,7 +28,7 @@ fun OnboardingView5(pagerState: PagerState) {
             Column(
         Modifier.background(Color.White)
     ) {
-        OnboardingHeader(pagerState = pagerState, title = "Favorites")
+        NavHeader(pagerState = pagerState, title = "Favorites")
 
         Column(
             modifier = Modifier
